refactor: split native wrapper responsibilities
