+++ src/native/index.ts
@@ -1,240 +1,59 @@
-import * as path from "path";
-import * as os from "os";
-import * as module from "module";
-import { fileURLToPath } from "url";
+import type {
+  BranchDelta,
+  CallEdgeData,
+  CallSiteData,
+  CentralityData,
+  ChunkData,
+  ChunkMetadata,
+  CodeChunk,
+  CommunityData,
+  DatabaseStats,
+  DynamicBatchOptions,
+  FileInput,
+  KeywordSearchResult,
+  ParsedFile,
+  ParsedSymbol,
+  PathHopData,
+  ReachabilityData,
+  SearchResult,
+  SymbolData,
+  CallType,
+  Confidence,
+  ChunkType,
+} from "./types.js";
+import { native } from "./binding.js";
+import {
+  createDynamicBatches,
+  createEmbeddingText,
+  createEmbeddingTexts,
+  estimateTokens,
+} from "./embedding.js";
 
-function getNativeBinding() {
-  const platform = os.platform();
-  const arch = os.arch();
+export type {
+  BranchDelta,
+  CallEdgeData,
+  CallSiteData,
+  CallType,
+  CentralityData,
+  ChunkData,
+  ChunkMetadata,
+  ChunkType,
+  Confidence,
+  CodeChunk,
+  CommunityData,
+  DatabaseStats,
+  DynamicBatchOptions,
+  FileInput,
+  KeywordSearchResult,
+  ParsedFile,
+  ParsedSymbol,
+  PathHopData,
+  ReachabilityData,
+  SearchResult,
+  SymbolData,
+};
 
-  let bindingName: string;
-  
-  if (platform === "darwin" && arch === "arm64") {
-    bindingName = "codebase-index-native.darwin-arm64.node";
-  } else if (platform === "darwin" && arch === "x64") {
-    bindingName = "codebase-index-native.darwin-x64.node";
-  } else if (platform === "linux" && arch === "x64") {
-    bindingName = "codebase-index-native.linux-x64-gnu.node";
-  } else if (platform === "linux" && arch === "arm64") {
-    bindingName = "codebase-index-native.linux-arm64-gnu.node";
-  } else if (platform === "win32" && arch === "x64") {
-    bindingName = "codebase-index-native.win32-x64-msvc.node";
-  } else {
-    throw new Error(`Unsupported platform: ${platform}-${arch}`);
-  }
-
-  // Determine the current directory - handle ESM, CJS, and bundled contexts
-  let currentDir: string;
-  let requireTarget: string;
-  
-  // Check for ESM context with valid import.meta.url
-  if (typeof import.meta !== 'undefined' && import.meta.url) {
-    currentDir = path.dirname(fileURLToPath(import.meta.url));
-    requireTarget = import.meta.url;
-  } 
-  // Fallback to __dirname for CJS/bundled contexts
-  else if (typeof __dirname !== 'undefined') {
-    currentDir = __dirname;
-    requireTarget = __filename;
-  }
-  // Last resort: use process.cwd() - shouldn't normally hit this
-  else {
-    currentDir = process.cwd();
-    requireTarget = path.join(currentDir, "index.js");
-  }
-  
-  // The native module is in the 'native' folder at package root
-  // From dist/index.js, we go up one level to package root, then into native/
-  // From src/native/index.ts (dev/test), we go up two levels to package root
-  const normalizedDir = currentDir.replace(/\\/g, '/');
-  const isDevMode = normalizedDir.includes('/src/native') || currentDir.includes(path.join('src', 'native'));
-  const packageRoot = isDevMode
-    ? path.resolve(currentDir, '../..')
-    : path.resolve(currentDir, '..');
-  const nativePath = path.join(packageRoot, 'native', bindingName);
-  
-  // Load the native module - use standard require for .node files
-  const require = module.createRequire(requireTarget);
-  return require(nativePath);
-}
-
-function createMockNativeBinding() {
-  const error = new Error("Native module not available. Please rebuild with 'npm run build:native'.");
-  
-  return {
-    parseFile: () => { throw error; },
-    parseFiles: () => { throw error; },
-    hashContent: () => { throw error; },
-    hashFile: () => { throw error; },
-    extractCalls: () => { throw error; },
-    VectorStore: class {
-      constructor() { throw error; }
-    },
-    InvertedIndex: class {
-      constructor() { throw error; }
-      serialize() { throw error; }
-      deserialize() { throw error; }
-    },
-    Database: class {
-      constructor() { throw error; }
-      static openReadOnly() { throw error; }
-      static createEmptyReadOnly() { throw error; }
-      close() { throw error; }
-      getTransitiveReachability() { throw error; }
-      detectCommunities() { throw error; }
-      computeCentrality() { throw error; }
-    },
-}
-}
-
-let native: any;
-try {
-  native = getNativeBinding();
-} catch (e) {
-  console.error("[codebase-index] Failed to load native module:", e);
-  native = createMockNativeBinding();
-}
-
-export interface FileInput {
-  path: string;
-  content: string;
-}
-
-export interface CodeChunk {
-  content: string;
-  startLine: number;
-  startCol?: number;
-  endLine: number;
-  endCol?: number;
-  chunkType: ChunkType;
-  name?: string;
-  language: string;
-}
-
-export type ChunkType =
-  | "function"
-  | "class"
-  | "method"
-  | "interface"
-  | "type"
-  | "enum"
-  | "struct"
-  | "impl"
-  | "trait"
-  | "module"
-  | "import"
-  | "export"
-  | "comment"
-  | "other";
-
-export interface ParsedFile {
-  path: string;
-  chunks: CodeChunk[];
-  symbols: ParsedSymbol[];
-  hash: string;
-}
-
-export interface ParsedSymbol {
-  name: string;
-  kind: string;
-  startLine: number;
-  startCol: number;
-  endLine: number;
-  endCol: number;
-  language: string;
-}
-
-
-export type Confidence = "Direct" | "Inferred";
-
-export type CallType = "Call" | "MethodCall" | "Constructor" | "Import" | "Inherits" | "Implements";
-
-export interface CallSiteData {
-  calleeName: string;
-  line: number;
-  column: number;
-  callType: CallType;
-  confidence: Confidence;
-}
-
-export interface SymbolData {
-  id: string;
-  filePath: string;
-  name: string;
-  kind: string;
-  startLine: number;
-  startCol: number;
-  endLine: number;
-  endCol: number;
-  language: string;
-}
-
-export interface CallEdgeData {
-  id: string;
-  fromSymbolId: string;
-  fromSymbolName?: string;
-  fromSymbolFilePath?: string;
-  targetName: string;
-  toSymbolId?: string;
-  callType: string;
-  confidence: string;
-  line: number;
-  col: number;
-  isResolved: boolean;
-}
-
-export interface PathHopData {
-  symbolId: string;
-  symbolName: string;
-  filePath: string;
-  line: number;
-  callType: string;
-}
-
-export interface ReachabilityData {
-  symbolId: string;
-  symbolName: string;
-  filePath: string;
-  depth: number;
-}
-
-export interface CommunityData {
-  symbolId: string;
-  symbolName: string;
-  filePath: string;
-  communityId: number;
-  communityLabel: string;
-}
-
-export interface CentralityData {
-  symbolId: string;
-  symbolName: string;
-  filePath: string;
-  callerCount: number;
-  calleeCount: number;
-  totalConnections: number;
-}
-
-export interface SearchResult {
-  id: string;
-  score: number;
-  metadata: ChunkMetadata;
-}
-
-export interface ChunkMetadata {
-  filePath: string;
-  startLine: number;
-  endLine: number;
-  chunkType: ChunkType;
-  name?: string;
-  language: string;
-  hash: string;
-  blameSha?: string;
-  blameAuthor?: string;
-  blameAuthorEmail?: string;
-  blameCommittedAt?: number;
-  blameSummary?: string;
-}
+export { estimateTokens, createDynamicBatches, createEmbeddingText, createEmbeddingTexts };
 
 export function parseFile(filePath: string, content: string): CodeChunk[] {
   const result = native.parseFile(filePath, content);
@@ -288,7 +107,6 @@ export function hashContent(content: string): string {
 export function hashFile(filePath: string): string {
   return native.hashFile(filePath);
 }
-
 
 export function extractCalls(content: string, language: string): CallSiteData[] {
   return native.extractCalls(content, language);
@@ -404,297 +222,6 @@ export class VectorStore {
   }
 }
 
-// Token estimation: ~4 chars per token for code (conservative)
-const CHARS_PER_TOKEN = 4;
-const MAX_BATCH_TOKENS = 7500; // Leave buffer under 8192 API limit
-const MAX_SINGLE_CHUNK_TOKENS = 2000; // Default truncation cap for individual chunks
-
-export function estimateTokens(text: string): number {
-  return Math.ceil(text.length / CHARS_PER_TOKEN);
-}
-
-function getEmbeddingHeaderParts(chunk: CodeChunk, filePath: string): string[] {
-  const parts: string[] = [];
-
-  const fileName = filePath.split("/").pop() || filePath;
-  const dirPath = filePath.split("/").slice(-3, -1).join("/");
-
-  const langDescriptors: Record<string, string> = {
-    typescript: "TypeScript",
-    javascript: "JavaScript",
-    python: "Python",
-    rust: "Rust",
-    swift: "Swift",
-    go: "Go",
-    java: "Java",
-  };
-
-  const typeDescriptors: Record<string, string> = {
-    function_declaration: "function",
-    function: "function",
-    arrow_function: "arrow function",
-    method_definition: "method",
-    method_declaration: "method",
-    protocol_function_declaration: "protocol requirement",
-    init_declaration: "initializer",
-    deinit_declaration: "deinitializer",
-    subscript_declaration: "subscript",
-    class_declaration: "class",
-    actor_declaration: "actor",
-    extension_declaration: "extension",
-    protocol_declaration: "protocol",
-    struct_declaration: "struct",
-    interface_declaration: "interface",
-    type_alias_declaration: "type alias",
-    enum_declaration: "enum",
-    export_statement: "export",
-    lexical_declaration: "variable declaration",
-    function_definition: "function",
-    class_definition: "class",
-    function_item: "function",
-    impl_item: "implementation",
-    struct_item: "struct",
-    enum_item: "enum",
-    trait_item: "trait",
-  };
-
-  const lang = langDescriptors[chunk.language] || chunk.language;
-  const typeDesc = typeDescriptors[chunk.chunkType] || chunk.chunkType;
-
-  if (chunk.name) {
-    parts.push(`${lang} ${typeDesc} "${chunk.name}"`);
-  } else {
-    parts.push(`${lang} ${typeDesc}`);
-  }
-
-  if (dirPath) {
-    parts.push(`in ${dirPath}/${fileName}`);
-  } else {
-    parts.push(`in ${fileName}`);
-  }
-
-  const semanticHints = extractSemanticHints(chunk.name || "", chunk.content);
-  if (semanticHints.length > 0) {
-    parts.push(`Purpose: ${semanticHints.join(", ")}`);
-  }
-
-  return parts;
-}
-
-function buildEmbeddingText(headerParts: string[], content: string, partIndex?: number, partCount?: number): string {
-  const parts = [...headerParts];
-  if (partCount && partCount > 1 && partIndex) {
-    parts.push(`Part ${partIndex}/${partCount}`);
-  }
-  parts.push("");
-  parts.push(content);
-  return parts.join("\n");
-}
-
-function splitOversizedContent(content: string, maxContentChars: number): string[] {
-  if (content.length <= maxContentChars) {
-    return [content];
-  }
-
-  const overlapChars = Math.max(CHARS_PER_TOKEN * 32, Math.min(Math.floor(maxContentChars * 0.15), CHARS_PER_TOKEN * 128));
-  const stepChars = Math.max(1, maxContentChars - overlapChars);
-  const segments: string[] = [];
-
-  for (let start = 0; start < content.length; start += stepChars) {
-    const end = Math.min(content.length, start + maxContentChars);
-    segments.push(content.slice(start, end));
-    if (end >= content.length) {
-      break;
-    }
-  }
-
-  return segments;
-}
-
-export function createEmbeddingTexts(chunk: CodeChunk, filePath: string, maxChunkTokens = MAX_SINGLE_CHUNK_TOKENS): string[] {
-  const headerParts = getEmbeddingHeaderParts(chunk, filePath);
-  const headerLength = buildEmbeddingText(headerParts, "", 1, 9).length;
-  const maxContentChars = Math.max(1, (maxChunkTokens * CHARS_PER_TOKEN) - headerLength);
-  const segments = splitOversizedContent(chunk.content, maxContentChars);
-
-  if (segments.length === 1) {
-    return [buildEmbeddingText(headerParts, segments[0])];
-  }
-
-  return segments.map((segment, index) => buildEmbeddingText(headerParts, segment, index + 1, segments.length));
-}
-
-export function createEmbeddingText(chunk: CodeChunk, filePath: string, maxChunkTokens = MAX_SINGLE_CHUNK_TOKENS): string {
-  const text = createEmbeddingTexts(chunk, filePath, maxChunkTokens)[0];
-  if (!text) {
-    return "";
-  }
-
-  const maxChars = maxChunkTokens * CHARS_PER_TOKEN;
-  if (text.length <= maxChars) {
-    return text;
-  }
-
-  return text.slice(0, Math.max(0, maxChars - 17)) + "\n... [truncated]";
-}
-
-export interface DynamicBatchOptions {
-  maxBatchTokens?: number;
-  maxBatchItems?: number;
-}
-
-export function createDynamicBatches<T extends { text: string; tokenCount?: number }>(chunks: T[], options: DynamicBatchOptions = {}): T[][] {
-  const batches: T[][] = [];
-  let currentBatch: T[] = [];
-  let currentTokens = 0;
-  const maxBatchTokens = Math.max(1, options.maxBatchTokens ?? MAX_BATCH_TOKENS);
-  const maxBatchItems = Math.max(1, options.maxBatchItems ?? Number.MAX_SAFE_INTEGER);
-  
-  for (const chunk of chunks) {
-    const chunkTokens = chunk.tokenCount ?? estimateTokens(chunk.text);
-
-    if (
-      currentBatch.length > 0
-      && (currentTokens + chunkTokens > maxBatchTokens || currentBatch.length >= maxBatchItems)
-    ) {
-      batches.push(currentBatch);
-      currentBatch = [];
-      currentTokens = 0;
-    }
-    
-    currentBatch.push(chunk);
-    currentTokens += chunkTokens;
-  }
-  
-  if (currentBatch.length > 0) {
-    batches.push(currentBatch);
-  }
-  
-  return batches;
-}
-
-function extractSemanticHints(name: string, content: string): string[] {
-  const hints: string[] = [];
-  const combined = `${name} ${content}`.toLowerCase();
-  
-  const signature = extractFunctionSignature(content);
-  if (signature) {
-    hints.push(signature);
-  }
-  
-  const patterns: Array<[RegExp, string]> = [
-    [/auth|login|logout|signin|signout|credential/i, "authentication"],
-    [/password|hash|bcrypt|argon/i, "password handling"],
-    [/token|jwt|bearer|oauth/i, "token management"],
-    [/user|account|profile|member/i, "user management"],
-    [/permission|role|access|authorize/i, "authorization"],
-    [/validate|verify|check|assert/i, "validation"],
-    [/error|exception|throw|catch/i, "error handling"],
-    [/log|debug|trace|info|warn/i, "logging"],
-    [/cache|memoize|store/i, "caching"],
-    [/fetch|request|response|api|http/i, "HTTP/API"],
-    [/database|db|query|sql|mongo/i, "database"],
-    [/file|read|write|stream|path/i, "file operations"],
-    [/parse|serialize|json|xml/i, "data parsing"],
-    [/encrypt|decrypt|crypto|secret|cipher|cryptographic/i, "encryption/cryptography"],
-    [/test|spec|mock|stub|expect/i, "testing"],
-    [/config|setting|option|env/i, "configuration"],
-    [/route|endpoint|handler|controller|middleware/i, "routing/middleware"],
-    [/render|component|view|template/i, "UI rendering"],
-    [/state|redux|store|dispatch/i, "state management"],
-    [/hook|effect|memo|callback/i, "React hooks"],
-  ];
-  
-  for (const [pattern, hint] of patterns) {
-    if (pattern.test(combined) && !hints.includes(hint)) {
-      hints.push(hint);
-    }
-  }
-  
-  return hints.slice(0, 6);
-}
-
-function extractFunctionSignature(content: string): string | null {
-  const tsJsPatterns = [
-    /(?:export\s+)?(?:async\s+)?function\s+(\w+)\s*(?:<[^>]+>)?\s*\(([^)]*)\)\s*(?::\s*([^{]+))?/,
-    /(?:export\s+)?const\s+(\w+)\s*(?::\s*[^=]+)?\s*=\s*(?:async\s+)?\(([^)]*)\)\s*(?::\s*([^=>{]+))?\s*=>/,
-    /(?:export\s+)?const\s+(\w+)\s*(?::\s*[^=]+)?\s*=\s*(?:async\s+)?function\s*\(([^)]*)\)/,
-  ];
-  
-  const pyPatterns = [
-    /def\s+(\w+)\s*\(([^)]*)\)\s*(?:->\s*([^:]+))?:/,
-    /async\s+def\s+(\w+)\s*\(([^)]*)\)\s*(?:->\s*([^:]+))?:/,
-  ];
-  
-  const goPatterns = [
-    /func\s+(?:\([^)]+\)\s+)?(\w+)\s*\(([^)]*)\)\s*(?:\(([^)]+)\)|([^{\n]+))?/,
-  ];
-  
-  const rustPatterns = [
-    /(?:pub\s+)?(?:async\s+)?fn\s+(\w+)\s*(?:<[^>]+>)?\s*\(([^)]*)\)\s*(?:->\s*([^{]+))?/,
-  ];
-  
-  for (const pattern of [...tsJsPatterns, ...pyPatterns, ...goPatterns, ...rustPatterns]) {
-    const match = content.match(pattern);
-    if (match) {
-      const funcName = match[1];
-      const params = match[2]?.trim() || "";
-      const returnType = (match[3] || match[4])?.trim();
-      
-      const paramNames = extractParamNames(params);
-      
-      let sig = `${funcName}(${paramNames.join(", ")})`;
-      if (returnType && returnType.length < 50) {
-        sig += ` -> ${returnType.replace(/\s+/g, " ").trim()}`;
-      }
-      
-      if (sig.length < 100) {
-        return sig;
-      }
-    }
-  }
-  
-  return null;
-}
-
-function extractParamNames(params: string): string[] {
-  if (!params.trim()) return [];
-  
-  const names: string[] = [];
-  const parts = params.split(",");
-  
-  for (const part of parts) {
-    const trimmed = part.trim();
-    if (!trimmed) continue;
-    
-    const tsMatch = trimmed.match(/^(\w+)\s*[?:]?/);
-    const pyMatch = trimmed.match(/^(\w+)\s*(?::|=)/);
-    const goMatch = trimmed.match(/^(\w+)\s+\w/);
-    const rustMatch = trimmed.match(/^(\w+)\s*:/);
-    
-    const match = tsMatch || pyMatch || goMatch || rustMatch;
-    if (match && match[1] !== "self" && match[1] !== "this") {
-      names.push(match[1]);
-    }
-  }
-  
-  return names.slice(0, 5);
-}
-
-export function generateChunkId(filePath: string, chunk: CodeChunk): string {
-  const hash = hashContent(`${filePath}:${chunk.startLine}:${chunk.endLine}:${chunk.content}`);
-  return `chunk_${hash.slice(0, 16)}`;
-}
-
-export function generateChunkHash(chunk: CodeChunk): string {
-  return hashContent(chunk.content);
-}
-
-export interface KeywordSearchResult {
-  chunkId: string;
-  score: number;
-}
-
 export class InvertedIndex {
   private inner: any;
 
@@ -746,36 +273,6 @@ export class InvertedIndex {
   getDocumentCount(): number {
     return this.inner.documentCount();
   }
-}
-
-export interface ChunkData {
-  chunkId: string;
-  contentHash: string;
-  filePath: string;
-  startLine: number;
-  endLine: number;
-  nodeType?: string;
-  name?: string;
-  language: string;
-  blameSha?: string;
-  blameAuthor?: string;
-  blameAuthorEmail?: string;
-  blameCommittedAt?: number;
-  blameSummary?: string;
-}
-
-export interface BranchDelta {
-  added: string[];
-  removed: string[];
-}
-
-export interface DatabaseStats {
-  embeddingCount: number;
-  chunkCount: number;
-  branchChunkCount: number;
-  branchCount: number;
-  symbolCount: number;
-  callEdgeCount: number;
 }
 
 export class Database {
@@ -994,8 +491,6 @@ export class Database {
     return this.inner.getStats();
   }
 
-
-
   upsertSymbol(symbol: SymbolData): void {
     this.throwIfClosed();
     this.inner.upsertSymbol(symbol);
@@ -1041,8 +536,6 @@ export class Database {
     return this.inner.deleteSymbolsByFile(filePath);
   }
 
-
-
   upsertCallEdge(edge: CallEdgeData): void {
     this.throwIfClosed();
     this.inner.upsertCallEdge(edge);
@@ -1059,7 +552,11 @@ export class Database {
     return this.inner.getCallers(targetName, branch, callTypeFilter ?? null);
   }
 
-  getCallersWithContext(targetName: string, branch: string, callTypeFilter?: string): CallEdgeData[] {
+  getCallersWithContext(
+    targetName: string,
+    branch: string,
+    callTypeFilter?: string
+  ): CallEdgeData[] {
     this.throwIfClosed();
     return this.inner.getCallersWithContext(targetName, branch, callTypeFilter ?? null);
   }
@@ -1079,12 +576,15 @@ export class Database {
     this.inner.resolveCallEdge(edgeId, toSymbolId);
   }
 
-  findShortestPath(fromName: string, toName: string, branch: string, maxDepth?: number): PathHopData[] {
+  findShortestPath(
+    fromName: string,
+    toName: string,
+    branch: string,
+    maxDepth?: number
+  ): PathHopData[] {
     this.throwIfClosed();
     return this.inner.findShortestPath(fromName, toName, branch, maxDepth ?? null);
   }
-
-
 
   addSymbolsToBranch(branch: string, symbolIds: string[]): void {
     this.throwIfClosed();
@@ -1125,8 +625,6 @@ export class Database {
     return this.inner.deleteBranchSymbolsForBranch(branch, symbolIds);
   }
 
-
-
   gcOrphanSymbols(): number {
     this.throwIfClosed();
     return this.inner.gcOrphanSymbols();
@@ -1164,4 +662,13 @@ export class Database {
     this.throwIfClosed();
     return this.inner.computeCentrality(branch);
   }
+}
+
+export function generateChunkId(filePath: string, chunk: CodeChunk): string {
+  const hash = hashContent(`${filePath}:${chunk.startLine}:${chunk.endLine}:${chunk.content}`);
+  return `chunk_${hash.slice(0, 16)}`;
+}
+
+export function generateChunkHash(chunk: CodeChunk): string {
+  return hashContent(chunk.content);
 }
